Worked on API
* Moved API classes to own package.
* Deleted IAntHill and IAnt in favour of BlockAntHill and ItemAnt.
* Finished javadocs for API classes.

+++ src/minecraft/vivadaylight3/myrmecology/common/lib/Environment.java
@@ -87,14 +87,14 @@ public class Environment {
 	if (!world.isRemote
 		&& world.getGameRules().getGameRuleBooleanValue("doTileDrops")) {
 	    float f = 0.7F;
-	    double d0 = (double) (world.rand.nextFloat() * f)
-		    + (double) (1.0F - f) * 0.5D;
-	    double d1 = (double) (world.rand.nextFloat() * f)
-		    + (double) (1.0F - f) * 0.5D;
-	    double d2 = (double) (world.rand.nextFloat() * f)
-		    + (double) (1.0F - f) * 0.5D;
-	    EntityItem entityitem = new EntityItem(world, (double) x + d0,
-		    (double) y + d1, (double) z + d2, item);
+	    double d0 = world.rand.nextFloat() * f
+		    + (1.0F - f) * 0.5D;
+	    double d1 = world.rand.nextFloat() * f
+		    + (1.0F - f) * 0.5D;
+	    double d2 = world.rand.nextFloat() * f
+		    + (1.0F - f) * 0.5D;
+	    EntityItem entityitem = new EntityItem(world, x + d0,
+		    y + d1, z + d2, item);
 	    entityitem.delayBeforeCanPickup = 10;
 	    world.spawnEntityInWorld(entityitem);
 	}
